Move JSI kind mapping into own file, add // TODO: JSIVALUECONVERSION comments
